RF: Handle sound from file/array in base sound class
