+++ psychopy/sound/backend_ptb.py
@@ -237,109 +237,17 @@ class SoundPTB(_SoundBase):
         self.loops = self._loopsRequested
         # start with the base class method
         _SoundBase.setSound(self, value, secs, octave, hamming, log)
-
-    def _setSndFromFile(self, filename):
-        # alias default names (so it always points to default.png)
-        if filename in ft.defaultStim:
-            filename = Path(prefs.paths['assets']) / ft.defaultStim[filename]
-        self.sndFile = f = sf.SoundFile(filename)
-        self.sourceType = 'file'
-        self.sampleRate = f.samplerate
-        if self.channels == -1:  # if channels was auto then set to file val
-            self.channels = f.channels
-        fileDuration = float(len(f)) / f.samplerate  # needed for duration?
-        # process start time
-        if self.startTime and self.startTime > 0:
-            startFrame = self.startTime * self.sampleRate
-            self.sndFile.seek(int(startFrame))
-            self.t = self.startTime
-        else:
-            self.t = 0
-        # process stop time
-        if self.stopTime and self.stopTime > 0:
-            requestedDur = self.stopTime - self.t
-            self.duration = min(requestedDur, fileDuration)
-        else:
-            self.duration = fileDuration - self.t
-        # can now calculate duration in frames
-        self.durationFrames = int(round(self.duration * self.sampleRate))
-        # are we preloading or streaming?
-        if self.preBuffer == 0:
-            # no buffer - stream from disk on each call to nextBlock
-            pass
-        elif self.preBuffer == -1:
-            # full pre-buffer. Load requested duration to memory
-            sndArr = self.sndFile.read(
-                frames=int(self.sampleRate * self.duration))
-            self.sndFile.close()
-            self._setSndFromArray(sndArr)
-        self._channelCheck(
-            self.sndArr)  # Check for fewer channels in stream vs data array
-
-    def _setSndFromArray(self, thisArray):
-
-        self.sndArr = np.asarray(thisArray).astype('float32')
-        if thisArray.ndim == 1:
-            self.sndArr.shape = [len(thisArray), 1]  # make 2D for broadcasting
-        if self.channels == 2 and self.sndArr.shape[1] == 1:  # mono -> stereo
-            self.sndArr = self.sndArr.repeat(2, axis=1)
-        elif self.sndArr.shape[1] == 1:  # if channels in [-1,1] then pass
-            pass
-        else:
-            try:
-                self.sndArr.shape = [len(thisArray), 2]
-            except ValueError:
-                raise ValueError("Failed to format sound with shape {} "
-                                 "into sound with channels={}"
-                                 .format(self.sndArr.shape, self.channels))
-
-        # is this stereo?
-        if self.stereo == -1:  # auto stereo. Try to detect
-            if self.sndArr.shape[1] == 1:
-                self.stereo = 0
-            elif self.sndArr.shape[1] == 2:
-                self.stereo = 1
-            else:
-                raise IOError("Couldn't determine whether array is "
-                              "stereo. Shape={}".format(self.sndArr.shape))
-        self._nSamples = thisArray.shape[0]
-        if self.stopTime == -1:
-            self.duration = self._nSamples / float(self.sampleRate)
-        # set to run from the start:
-        self.seek(0)
-        self.sourceType = "array"
-
-        # catch when array is empty
-        if not len(self.sndArr):
-            logging.warning(
-                "Received a blank array for sound, playing nothing instead."
-            )
-            self.sndArr = np.zeros(shape=(self.blockSize, self.channels))
-
-        # create audio clip
-        clip = AudioClip(
-            samples=self.sndArr, 
-            sampleRateHz=self.sampleRate
-        )
-        # set from clip
-        self._setSndFromClip(clip)
     
     def _setSndFromClip(self, clip: AudioClip):
-        """
-        Set current sound from an AudioClip object. All other setSound methods eventually lead to 
-        this - they just transform the given sound to an AudioClip.
-
-        Parameters
-        ----------
-        clip : AudioClip
-            AudioClip object to set.
-        """
         # store clip
         self.clip = clip
         # resample the clip if needed and allowed
         if self.speaker.resample:
             if clip.sampleRateHz != self.speaker.sampleRateHz:
                 clip.resample(targetSampleRateHz=self.speaker.sampleRateHz)
+        # work out stop time
+        if self.stopTime == -1:
+            self.duration = clip.sample.shape[0] / clip.sampleRateHz
         # create/update track
         if  self.track:
             self.track.stop()
@@ -350,6 +258,8 @@ class SoundPTB(_SoundBase):
                 data=clip.samples,
                 volume=self.volume
             )
+        # seek to start
+        self.seek(0)
 
     def _channelCheck(self, array):
         """Checks whether stream has fewer channels than data. If True, ValueError"""
